+++ src/commands/move.ts
@@ -32,13 +32,19 @@ export class ForwardChar extends EmacsCommand {
   public readonly id = "forwardChar";
 
   public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined) {
-    if (prefixArgument === undefined || prefixArgument === 1) {
+    const charDelta = prefixArgument == undefined ? 1 : prefixArgument;
+    if (this.emacsController.inRectMarkMode) {
+      this.emacsController.moveRectActives((curActives) => curActives.map((a) => a.translate(0, charDelta)));
+      return;
+    }
+
+    if (charDelta === 1) {
       return vscode.commands.executeCommand<void>(isInMarkMode ? "cursorRightSelect" : "cursorRight");
-    } else if (prefixArgument > 0) {
+    } else if (charDelta > 0) {
       const doc = textEditor.document;
       const newSelections = textEditor.selections.map((selection) => {
         const offset = doc.offsetAt(selection.active);
-        const newActivePos = doc.positionAt(offset + prefixArgument);
+        const newActivePos = doc.positionAt(offset + charDelta);
         const newAnchorPos = isInMarkMode ? selection.anchor : newActivePos;
         return new vscode.Selection(newAnchorPos, newActivePos);
       });
@@ -52,13 +58,21 @@ export class BackwardChar extends EmacsCommand {
   public readonly id = "backwardChar";
 
   public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined) {
-    if (prefixArgument === undefined || prefixArgument === 1) {
+    const charDelta = prefixArgument == undefined ? 1 : prefixArgument;
+    if (this.emacsController.inRectMarkMode) {
+      this.emacsController.moveRectActives((curActives) =>
+        curActives.map((a) => new vscode.Position(a.line, Math.max(a.character - charDelta, 0)))
+      );
+      return;
+    }
+
+    if (charDelta === 1) {
       return vscode.commands.executeCommand<void>(isInMarkMode ? "cursorLeftSelect" : "cursorLeft");
-    } else if (prefixArgument > 0) {
+    } else if (charDelta > 0) {
       const doc = textEditor.document;
       const newSelections = textEditor.selections.map((selection) => {
         const offset = doc.offsetAt(selection.active);
-        const newActivePos = doc.positionAt(offset - prefixArgument);
+        const newActivePos = doc.positionAt(offset - charDelta);
         const newAnchorPos = isInMarkMode ? selection.anchor : newActivePos;
         return new vscode.Selection(newAnchorPos, newActivePos);
       });
@@ -72,12 +86,20 @@ export class NextLine extends EmacsCommand {
   public readonly id = "nextLine";
 
   public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined) {
-    const value = prefixArgument === undefined ? 1 : prefixArgument;
+    const lineDelta = prefixArgument == undefined ? 1 : prefixArgument;
+
+    if (this.emacsController.inRectMarkMode) {
+      const maxLine = textEditor.document.lineCount - 1;
+      this.emacsController.moveRectActives((curActives) =>
+        curActives.map((a) => new vscode.Position(Math.min(a.line + lineDelta, maxLine), a.character))
+      );
+      return;
+    }
 
     return vscode.commands.executeCommand<void>("cursorMove", {
       to: "down",
       by: "wrappedLine",
-      value,
+      value: lineDelta,
       select: isInMarkMode,
     });
   }
@@ -87,12 +109,19 @@ export class PreviousLine extends EmacsCommand {
   public readonly id = "previousLine";
 
   public execute(textEditor: TextEditor, isInMarkMode: boolean, prefixArgument: number | undefined) {
-    const value = prefixArgument === undefined ? 1 : prefixArgument;
+    const lineDelta = prefixArgument == undefined ? 1 : prefixArgument;
+
+    if (this.emacsController.inRectMarkMode) {
+      this.emacsController.moveRectActives((curActives) =>
+        curActives.map((a) => new vscode.Position(Math.max(a.line - lineDelta, 0), a.character))
+      );
+      return;
+    }
 
     return vscode.commands.executeCommand<void>("cursorMove", {
       to: "up",
       by: "wrappedLine",
-      value,
+      value: lineDelta,
       select: isInMarkMode,
     });
   }