+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to the "emacs-mcx" extension will be documented in this file.
 
+## [Unreleased]
+### Add
+- (Experimental) `C-x <SPC>` (rectangle-mark-mode), #819.
+
 ## [0.31.0] - 2021-04-20
 ### Fixed
 - Use `isearchExit` with `then` arg instead of `executeCommands` to quit the find widget with key press.